requirements.txt
anthropic>=0.40.0
openai>=1.50.0
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
boto3>=1.35.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
alembic>=1.13.0
python-dotenv>=1.0.0
pydantic>=2.9.0
httpx>=0.27.0
streamlit>=1.38.0
pytest>=8.3.0
pytest-asyncio>=0.24.0
requests>=2.32.0
python-multipart>=0.0.9
python-jose[cryptography]>=3.3.0
pdf2image>=1.17.0
